[carthage] add - documentation
